// src/app/personal/interfaces/children.interface.ts
// Generated by https://quicktype.io

export interface Children {
    id:         string;
    tipoDoc:    TipoDoc;
    numeroDoc:  string;
    apPaterno?:  string;
    apMaterno?:  string;
    nombre1?:    string;
    nombre2?:    string;
    nombreCompleto?:    string;
    fechaNac:   string;
    idPersonal: string;
}


export enum TipoDoc {
    Dni = "DNI",
    Ruc = "RUC",
}